refactor : ♻️ 상수 파일 분리
