+++ week2/assignment/가계부 💸/index.html
@@ -46,7 +46,7 @@
     <footer>
         <button id="add">내역 추가</button>
     </footer>
-    <script src="js/initData.js"></script>
+    <script type="module" src="js/initData.js"></script>
 </body>
 
 </html>
